APNS: We now keep both services present, production & sandbox. For this reason we also actually will have two ApnsServices in the future connecting to the different services at apple.
To be clarified: It seems we will also actually require two certificates for that in the future and not only one. Furthermore, we need to determine which service to actually use for which client (will be based on information transmitted by the client during hello()).

I have no idea how to test all that in the end...