Changed the way we persist to mongo db to the same paradigm as all other models.
The update nice and dandy but leads to potentially fields not being removed if the model changes. Also unclear what actually happens to fields that are set to a value but get nulled via update and vice versa.

The new (old) way ensures that the database has a proper reflection of what actually was received.