Optimizing the performance for the migration execution.
Before optimization it was dog slow. I do not think that migrations are the way to go in the future. We should probably do things like this on-the-fly in the future.

The mechanism could be something like e.g. in the JongoDatabase guard all return values and send them through some piece of code that guarantees their validity. Migrations in the way we currently have would not work. We might, though, mark every entry like e.g. TalkDelivery with a version.

Definitely needs further discussion.